Redo CR suggestions
